+++ util.h
@@ -8,7 +8,7 @@ void tile2latlon(unsigned int x, unsigned int y, int zoom, double *lat, double *
 void wxy2fxy(long long wx, long long wy, double *ox, double *oy, int z, int x, int y);
 void xy2buf(unsigned int x32, unsigned int y32, unsigned char *buf, int *offbits, int n, int skip);
 void zxy2bufs(unsigned int z, unsigned int x, unsigned int y, unsigned char *startbuf, unsigned char *endbuf, int bytes);
-void buf2xys(unsigned char *buf, int mapbits, int metabits, int skip, int n, unsigned int *x, unsigned int *y, unsigned int *meta);
+void buf2xys(unsigned char *buf, int mapbits, int metabits, int skip, int n, unsigned int *x, unsigned int *y, unsigned long long *meta);
 void meta2buf(int bits, long long data, unsigned char *buf, int *offbits, int max);
 
 int bytesfor(int mapbits, int metabits, int components, int z_lookup);
